fix: address pr suggestion
